requirements.txt
langchain
langchain-community
langchain-openai
faiss-cpu
pypdf
chromadb
sentence-transformers
streamlit
openai
python-dotenv
pandas
altair
